munnir-api/pyproject.toml
[project]
name = "munnir-api"
version = "0.1.0"
description = "Munnir geopolitical trading simulator — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "email-validator>=2.0",
    "pybind11>=2.13",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "httpx>=0.28",
    "google-genai>=1.0",
    "feedparser>=6.0",
    "tenacity>=9.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
